requirements.txt
aiohttp==3.12.1
fastapi==0.115.12
pydantic==2.11.5
python-dotenv==1.1.0
